Simulation based density for visual inference

Probablity of observing exactly x picks of the data plot in K evaluations of a lineup of size m. We distinguish between three different scenarios:

Scenario I: in each of K evaluations a different data set and a different set of (m-1) null plots is shown.

Scenario II: in each of K evaluations the same data set but a different set of (m-1) null plots is shown.

Scenario III: the same lineup, i.e. same data and same set of null plots, is shown to K different observers.

dVsim(x, K, m = 20, N = 10000, scenario = 3, xp = 1, target = 1)

Arguments

x
number of observed picks of the data plot

K
number of evaluations of the same lineup

m
size of the lineup

N
MC parameter: number of replicates on which MC probabilities are based. Higher number of replicates will decrease MC variability.

scenario
numeric value, one of 1, 2, or 3, indicating the type of simulation used: scenario 3 assumes that the same lineup is shown in all K evaluations

xp
exponent used, defaults to 1

target
location of target plot(s). By default 1. If several targets are present, specify vector of target locations.

Value

simulation based density to observe x picks of the data plot in K evaluation under the assumption that the data plot is consistent with the null hypothesis. For comparison a p value based on a binomial distribution is provided as well.

Examples

dVsim(2, 20, m=3) # triangle test

#>   x scenario3      binom
#> 3 2    0.0656 0.01428461

# NOT RUN {
## points in red are binomial distribution, black points are for inference
## in lineups using scenario 3
require(ggplot2)
qplot(x=x, y=scenario3, data=dVsim(0:6,6,m=2)) +
   geom_point(aes(x,y=binom), colour="red") + ylim(c(0,0.5))
qplot(x=x, y=scenario3, data=dVsim(0:6,6,m=3)) +
   geom_point(aes(x,y=binom), colour="red") + ylim(c(0,0.5))
# }

# lineup with two targets: what are the probabilities to identify at least 
# one of the targets?
dVsim(0:5, K=5, m=20, N=10000, scenario=3, target=1:2)

#>   x scenario3   binom
#> 1 0    0.5895 0.59049
#> 2 1    0.3173 0.32805
#> 3 2    0.0787 0.07290
#> 4 3    0.0141 0.00810
#> 5 4    0.0004 0.00045
#> 6 5    0.0000 0.00001

# slight difference between this distribution and the distribution for a 
# lineup of size 10 with a single target:
dVsim(0:5, K=5, m=10, N=10000, scenario=3, target=1)

#>   x scenario3   binom
#> 1 0    0.6137 0.59049
#> 2 1    0.2869 0.32805
#> 3 2    0.0861 0.07290
#> 4 3    0.0119 0.00810
#> 5 4    0.0012 0.00045
#> 6 5    0.0002 0.00001