Bootstrap based p values for visual inference

Simulation based p value to observe x or more picks of the data plot in K evaluations under the assumption that the data plot is consistent with the null hypothesis. We distinguish between three different scenarios:

Scenario I: in each of K evaluations a different data set and a different set of (m-1) null plots is shown.

Scenario II: in each of K evaluations the same data set but a different set of (m-1) null plots is shown.

Scenario III: the same lineup, i.e. same data and same set of null plots, is shown to K different observers.

pVsim(
  x,
  K,
  m = 20,
  N = 10000,
  scenario = 3,
  xp = 1,
  target = 1,
  upper.tail = TRUE
)

Arguments

x
number of observed picks of the data plot

K
number of evaluations

m
size of the lineup

N
MC parameter: number of replicates on which MC probabilities are based. Higher number of replicates will decrease MC variability.

scenario
numeric value, one of 1,2, or 3, indication the type of simulation used: scenario 3 assumes that the same lineup is shown in all K evaluations

xp
exponent used, defaults to 1

target
(vector) of integer values between 1 and m indicating the position(s) of the target plots. Only the number of targets will affect the probabilities.

upper.tail
compute probabilities P(X >= x). Be aware that the use of this parameter is not consistent with the other distribution functions in base. There, a value of P(X > x) is computed for upper.tail=TRUE.

Value

Vector/data frame. For comparison a p value based on a binomial distribution is provided as well.

Examples

pVsim(15, 20, m=3) # triangle test

#> Warning: `pVsim()` is deprecated as of vinference 1.0.0.
#> Please use `new()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#>       x simulated       binom
#> [1,] 15    0.0369 0.000167366