Theoretical quantile for lineups under different scenarios

Some more details to be written later

qV(q, K, m, scenario, type = "numeric")

Arguments

q
(vector) of the quantiles for picks of the data plot out of K evaluations

K
number of evaluations

m
size of the lineup

scenario
which scenario should be used? 1, 2, or 3?

type
one of "mpfr" or "numeric". Should the result be in arbitrary numeric length or be a numeric? Internally the Rmpfr package is used to get maximal precision.

Examples

## get critical values of visual triangle test:
qV(q=c(0.95, 0.99), K=c(5,10,15,20, 25, 30), m=3, scenario=1:3)

#>       q  K scenario1 scenario2 scenario3
#> 1  0.95 10         6         8         8
#> 2  0.95 15         8        10        12
#> 3  0.95 20        10        13        15
#> 4  0.95 25        12        16        18
#> 5  0.95 30        14        18        21
#> 6  0.95  5         3         5         5
#> 7  0.99 10         7         9        10
#> 8  0.99 15         9        12        14
#> 9  0.99 20        12        15        18
#> 10 0.99 25        14        18        22
#> 11 0.99 30        16        21        26
#> 12 0.99  5         4        NA        NA

## get critical values of full lineup test:
qV(q=c(0.95, 0.99), K=c(5,10,15,20, 25, 30), m=20, scenario=3)

#>       q  K scenario3
#> 1  0.95 10         3
#> 2  0.95 15         4
#> 3  0.95 20         4
#> 4  0.95 25         5
#> 5  0.95 30         5
#> 6  0.95  5         2
#> 7  0.99 10         4
#> 8  0.99 15         5
#> 9  0.99 20         5
#> 10 0.99 25         6
#> 11 0.99 30         7
#> 12 0.99  5         3